requirements.txt
Flask==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==2.1.5
Werkzeug==3.0.4
python-dotenv==1.0.1
spotipy==2.24.0
gunicorn==20.1.0
wheel
pyaudio==0.2.11
spacy>=3.0.0
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.0.0/es_core_news_sm-3.0.0.tar.gz
